Renamed type files and package scripts
