+++ package.json
@@ -1,35 +1,35 @@
 {
-  "name": "moderation-bot",
-  "version": "1.0.0",
-  "description": "The Roblox Discord's Moderation Bot",
-  "main": "src/client.ts",
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/Client.js",
-    "test": "ts-node src/client.ts"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Rodis-Infrastructure/moderation-bot.git"
-  },
-  "author": "",
-  "bugs": {
-    "url": "https://github.com/Rodis-Infrastructure/moderation-bot/issues"
-  },
-  "homepage": "https://github.com/Rodis-Infrastructure/moderation-bot#readme",
-  "dependencies": {
-    "@iarna/toml": "^2.2.5",
-    "discord.js": "^14.9.0",
-    "dotenv": "^16.0.3",
-    "ms": "^2.1.3",
-    "sqlite3": "^5.1.6"
-  },
-  "devDependencies": {
-    "@types/ms": "^0.7.31",
-    "@typescript-eslint/eslint-plugin": "^5.59.1",
-    "@typescript-eslint/parser": "^5.59.1",
-    "eslint": "^8.39.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4"
-  }
+    "name": "moderation-bot",
+    "version": "1.0.0",
+    "description": "The Roblox Discord's Moderation Bot",
+    "main": "src/client.ts",
+    "scripts": {
+        "build": "tsc",
+        "start": "node dist/client.js",
+        "dev": "ts-node src/client.ts"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/Rodis-Infrastructure/moderation-bot.git"
+    },
+    "author": "",
+    "bugs": {
+        "url": "https://github.com/Rodis-Infrastructure/moderation-bot/issues"
+    },
+    "homepage": "https://github.com/Rodis-Infrastructure/moderation-bot#readme",
+    "dependencies": {
+        "@iarna/toml": "^2.2.5",
+        "discord.js": "^14.9.0",
+        "dotenv": "^16.0.3",
+        "ms": "^2.1.3",
+        "sqlite3": "^5.1.6"
+    },
+    "devDependencies": {
+        "@types/ms": "^0.7.31",
+        "@typescript-eslint/eslint-plugin": "^5.59.1",
+        "@typescript-eslint/parser": "^5.59.1",
+        "eslint": "^8.39.0",
+        "ts-node": "^10.9.1",
+        "typescript": "^5.0.4"
+    }
 }
