+++ src/interactions/commands/kick.ts
@@ -7,7 +7,7 @@ import {
 
 import { resolveInfraction, validateModerationAction } from "../../utils/moderation";
 import { InteractionResponseType } from "../../types/interactions";
-import { InfractionPunishment } from "../../types/database";
+import { InfractionPunishment } from "../../types/db";
 import { formatReason } from "../../utils";
 
 import ChatInputCommand from "../../handlers/interactions/commands/chatInputCommand";
