+++ src/types/interactions.ts
